# apps/backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/quizzes": {
      "get": {
        "operationId": "QuizzesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "post": {
        "operationId": "QuizzesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/samples": {
      "post": {
        "operationId": "QuizzesController_createSamples",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/duplicate": {
      "post": {
        "operationId": "QuizzesController_duplicate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}": {
      "get": {
        "operationId": "QuizzesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "put": {
        "operationId": "QuizzesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "delete": {
        "operationId": "QuizzesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/feedback": {
      "get": {
        "operationId": "QuizzesController_feedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "rating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizFeedbackSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/sessions": {
      "get": {
        "operationId": "QuizzesController_sessions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/sessions/{sessionId}": {
      "get": {
        "operationId": "QuizzesController_sessionDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/sessions/{sessionId}/players/{playerResultId}": {
      "get": {
        "operationId": "QuizzesController_sessionPlayer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerResultId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionPlayerDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/status": {
      "patch": {
        "operationId": "QuizzesController_transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
...
  },
  "info": {
    "title": "QuizDock API",
    "description": "API REST du builder de quiz et des restitutions",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "QuizDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "coverMediaId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "ready",
              "archived"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "unlisted"
            ]
          },
          "language": {
            "type": "string"
          },
          "feedbackEnabled": {
            "type": "boolean"
          },
          "questionCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "ownerId",
          "title",
          "description",
          "coverMediaId",
          "status",
          "visibility",
          "language",
          "feedbackEnabled",
          "questionCount",
          "createdAt",
          "updatedAt",
          "archivedAt"
        ]
      },
      "CreateQuizDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "language": {
            "default": "fr",
            "type": "string",
            "minLength": 2,
            "maxLength": 5
          },
          "feedbackEnabled": {
            "type": "boolean"
          },
          "coverMediaId": {
            "type": "string",
            "minLength": 26,
            "maxLength": 26
          }
        },
        "required": [
          "title"
        ]
      },
...
      "QuizFeedbackSummaryDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "average": {
            "type": "number"
          },
          "distribution": {
            "minItems": 5,
            "maxItems": 5,
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "rating": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "comment": {
                  "type": "string",
                  "nullable": true
                },
                "nickname": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "rating",
                "comment",
                "nickname",
                "createdAt"
              ]
            }
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "count",
          "average",
          "distribution",
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "SessionListDto": {
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "pin": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "lobby",
                    "in_progress",
                    "ended",
                    "interrupted",
                    "archived"
                  ]
                },
                "playerCount": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "successRate": {
                  "type": "number",
                  "nullable": true
                },
                "fullCapture": {
                  "type": "boolean"
                },
                "startedAt": {
                  "type": "string"
                },
                "endedAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "pin",
                "status",
                "playerCount",
                "successRate",
                "fullCapture",
                "startedAt",
                "endedAt"
              ]
            }
          }
        },
        "required": [
          "sessions"
        ]
      },
      "SessionDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pin": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "lobby",
              "in_progress",
              "ended",
              "interrupted",
              "archived"
            ]
          },
          "playerCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "successRate": {
            "type": "number",
            "nullable": true
          },
          "fullCapture": {
            "type": "boolean"
          },
          "startedAt": {
            "type": "string"
          },
          "endedAt": {
            "type": "string"
          },
          "quizTitle": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "totalQuestions": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "questions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "orderIndex": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "prompt": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "answerCount": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "correctCount": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "successRate": {
                  "type": "number"
                },
                "avgResponseMs": {
                  "type": "number",
                  "nullable": true
                }
              },
              "required": [
                "orderIndex",
                "prompt",
                "type",
                "answerCount",
                "correctCount",
                "successRate",
                "avgResponseMs"
              ]
            }
          },
          "players": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "nickname": {
                  "type": "string"
                },
                "finalRank": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "finalScore": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "correctCount": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "answeredCount": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "avgResponseMs": {
                  "type": "number",
                  "nullable": true
                },
                "maxStreak": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "id",
                "nickname",
                "finalRank",
                "finalScore",
                "correctCount",
                "answeredCount",
                "avgResponseMs",
                "maxStreak"
              ]
            }
          }
        },
        "required": [
          "id",
          "pin",
          "status",
          "playerCount",
          "successRate",
          "fullCapture",
          "startedAt",
          "endedAt",
          "quizTitle",
          "language",
          "totalQuestions",
          "questions",
          "players"
        ]
      },
      "SessionPlayerDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "finalRank": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "finalScore": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "correctCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "answeredCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "avgResponseMs": {
            "type": "number",
            "nullable": true
          },
          "maxStreak": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "fullCapture": {
            "type": "boolean"
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "orderIndex": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "prompt": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "answer": {
                  "type": "string"
                },
                "isCorrect": {
                  "type": "boolean"
                },
                "pointsAwarded": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "responseMs": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "orderIndex",
                "prompt",
                "type",
                "answer",
                "isCorrect",
                "pointsAwarded",
                "responseMs"
              ]
            }
          }
        },
        "required": [
          "id",
          "nickname",
          "finalRank",
          "finalScore",
          "correctCount",
          "answeredCount",
          "avgResponseMs",
          "maxStreak",
          "fullCapture",
          "answers"
        ]
      },
      "UpdateQuizDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 5
          },
          "feedbackEnabled": {
            "type": "boolean"
          },
          "coverMediaId": {
            "type": "string",
            "minLength": 26,
            "maxLength": 26,
            "nullable": true
          }
        }
      },
      "TransitionQuizDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "ready",
              "archived"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}